Move messages.py to constants
