+++ src/application/telegram_bot/constants/messages.py
